services/recommendation-service/requirements.txt
flask
SQLAlchemy
flask_sqlalchemy
pandas
numpy
scipy
surprise
cryptography
pyjwt
pymysql
scikit-learn